refactor: adopt domain-based packaging for zeedb
